+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#1539](https://github.com/bbatsov/projectile/pull/1539): New defcustom `projectile-auto-discover` controlling whether to automatically discover projects in the search path when `projectile-mode` activates.
+* Add [emacs-eldev](https://github.com/doublep/eldev) project type.
 
 ### Changes